refactor(frontend): split useMCPTool.ts (1536 lines) into focused modules
Break the monolithic hook file into 5 files with clear responsibilities:
- hooks/chat/tool-labels.ts — single source of truth for TOOL_LABELS
- hooks/chat/chat-helpers.ts — buildContextualQuery, extractNormsFromAnswer
- hooks/chat/evidence-extractor.ts — extractEvidenceFromToolResult (9 categories)
- hooks/chat/useAIChat.ts — agentic chat hook with two-phase plan flow
- hooks/useMCPTool.ts — manual tool hook + re-exports for backward compat

Also removes duplicated TOOL_LABELS from CostSummary and PlanReviewDisplay.
