+++ lexwebapp/src/hooks/chat/tool-labels.ts
@@ -1,0 +1,61 @@
+/**
+ * Canonical tool name labels for UI display.
+ * Single source of truth — imported by useMCPTool, useAIChat, CostSummary, PlanReviewDisplay.
+ */
+
+export const TOOL_LABELS: Record<string, string> = {
+  search_legal_precedents: 'Пошук прецедентів',
+  search_supreme_court_practice: 'Практика ВС',
+  get_court_decision: 'Отримання рішення',
+  get_case_documents_chain: 'Ланцюг документів',
+  find_similar_fact_pattern_cases: 'Схожі справи',
+  compare_practice_pro_contra: 'Аналіз за і проти',
+  count_cases_by_party: 'Справи сторони',
+  get_case_text: 'Текст справи',
+  analyze_case_pattern: 'Аналіз патерну',
+  get_similar_reasoning: 'Схоже обґрунтування',
+  get_citation_graph: 'Граф цитувань',
+  check_precedent_status: 'Статус прецеденту',
+  load_full_texts: 'Повні тексти рішень',
+  search_legislation: 'Пошук законодавства',
+  get_legislation_article: 'Стаття закону',
+  get_legislation_articles: 'Статті закону',
+  get_legislation_section: 'Розділ закону',
+  get_legislation_structure: 'Структура закону',
+  search_procedural_norms: 'Процесуальні норми',
+  find_relevant_law_articles: 'Релевантні статті',
+  store_document: 'Збереження документу',
+  list_documents: 'Список документів',
+  semantic_search: 'Семантичний пошук',
+  get_document: 'Отримання документу',
+  parse_document: 'Розбір документу',
+  extract_document_sections: 'Секції документу',
+  summarize_document: 'Резюме документу',
+  compare_documents: 'Порівняння документів',
+  extract_key_clauses: 'Ключові положення',
+  calculate_procedural_deadlines: 'Розрахунок строків',
+  build_procedural_checklist: 'Процесуальний чеклист',
+  calculate_monetary_claims: 'Грошові вимоги',
+  generate_dd_report: 'DD звіт',
+  risk_scoring: 'Скоринг ризиків',
+  format_answer_pack: 'Пакет відповідей',
+  rada_search_parliament_bills: 'Законопроекти Ради',
+  rada_get_deputy_info: 'Інфо про депутата',
+  rada_search_legislation_text: 'Текст законів',
+  rada_analyze_voting_record: 'Голосування',
+  openreyestr_search_entities: 'Пошук юросіб',
+  openreyestr_get_entity_details: 'Деталі юрособи',
+  openreyestr_search_beneficiaries: 'Пошук бенефіціарів',
+  openreyestr_get_by_edrpou: 'Пошук за ЄДРПОУ',
+  openreyestr_get_statistics: 'Статистика реєстру',
+  openreyestr_search_enforcement_proceedings: 'Виконавчі провадження',
+  openreyestr_search_debtors: 'Боржники',
+  openreyestr_search_bankruptcy_cases: 'Банкрутство',
+  openreyestr_search_notaries: 'Нотаріуси',
+  openreyestr_search_court_experts: 'Судові експерти',
+  openreyestr_search_arbitration_managers: 'Арбітражні керуючі',
+};
+
+export function getToolLabel(toolName: string): string {
+  return TOOL_LABELS[toolName] || toolName;
+}
